+++ src/app/key_bind.rs
@@ -1,0 +1,31 @@
+use std::collections::HashMap;
+
+use cosmic::widget::menu::key_bind::KeyBind;
+use cosmic::widget::menu::key_bind::Modifier;
+use cosmic::{iced::keyboard::Key, iced_core::keyboard::key::Named};
+
+use crate::app::Action;
+
+pub fn key_binds() -> HashMap<KeyBind, Action> {
+    let mut key_binds = HashMap::new();
+
+    macro_rules! bind {
+        ([$($modifier:ident),* $(,)?], $key:expr, $action:ident) => {{
+            key_binds.insert(
+                KeyBind {
+                    modifiers: vec![$(Modifier::$modifier),*],
+                    key: $key,
+                },
+                Action::$action,
+            );
+        }};
+    }
+    bind!([Ctrl], Key::Character("r".into()), CreateRepository);
+    bind!([Ctrl, Shift], Key::Character("r".into()), CreateSnapshot);
+    bind!([Ctrl], Key::Character("w".into()), WindowClose);
+    bind!([Ctrl, Shift], Key::Character("n".into()), WindowNew);
+    bind!([Ctrl], Key::Character(",".into()), Settings);
+    bind!([Ctrl], Key::Character("i".into()), About);
+
+    key_binds
+}
